+++ src/Products/PloneMeeting/browser/overrides.py
@@ -1395,7 +1395,10 @@ class PMContentHistoryView(IHContentHistoryView):
       Overrides the ContentHistoryView template to use our own.
       We want to display the content_history as a table.
     '''
-    histories_to_handle = (u'revision', u'workflow', u'data_changes', u'deleted_children')
+    histories_to_handle = (u'revision',
+                           u'workflow',
+                           u'data_changes',
+                           u'deleted_children')
 
     def show_history(self):
         """Override to take MeetingConfig.hideHistoryTo into account."""
@@ -1424,7 +1427,10 @@ class PMContentHistoryView(IHContentHistoryView):
 
 class AdviceContentHistoryView(PMContentHistoryView):
     """ """
-    histories_to_handle = (u'revision', u'workflow', u'advice_given')
+    histories_to_handle = (u'revision',
+                           u'workflow',
+                           u'advice_given',
+                           u'advice_hide_during_redaction')
 
     def show_preview(self, event):
         """ """