+++ src/Products/PloneMeeting/events.py
@@ -24,6 +24,7 @@ from imio.helpers.content import safe_delattr
 from imio.helpers.security import fplog
 from imio.helpers.workflow import update_role_mappings_for
 from imio.helpers.xhtml import storeImagesLocally
+from imio.history.utils import add_event_to_history
 from OFS.interfaces import IObjectWillBeAddedEvent
 from OFS.ObjectManager import BeforeDeleteException
 from persistent.list import PersistentList
@@ -930,21 +931,36 @@ def onAdviceModified(advice, event):
     if advice.REQUEST.get('currentlyStoringExternalImages', False) is True:
         return
 
-    # update advice_row_id
-    advice._updateAdviceRowId()
+    if not isinstance(event, ContainerModifiedEvent):
+        mod_attrs = get_modified_attrs(event)
+        if 'advice_hide_during_redaction' in mod_attrs:
+            # add a line to the advice's wf history
+            action = 'to_hidden_during_redaction_action'
+            comments = 'to_hidden_during_redaction_comments'
+            if advice.advice_hide_during_redaction is False:
+                action = 'to_not_hidden_during_redaction_action'
+                comments = 'to_not_hidden_during_redaction_comments'
+            add_event_to_history(
+                advice,
+                'advice_hide_during_redaction_history',
+                action=action,
+                comments=comments)
 
-    item = advice.getParentNode()
-    item.update_local_roles()
+        # update advice_row_id
+        advice._updateAdviceRowId()
 
-    # make sure external images used in RichText fields are stored locally
-    storeImagesLocallyDexterity(advice)
+        item = advice.getParentNode()
+        item.update_local_roles()
 
-    # notify our own PM event so we are sure that this event is called
-    # after the onAviceModified event
-    notify(AdviceAfterModifyEvent(advice))
+        # make sure external images used in RichText fields are stored locally
+        storeImagesLocallyDexterity(advice)
 
-    # update item
-    _advice_update_item(item)
+        # notify our own PM event so we are sure that this event is called
+        # after the onAviceModified event
+        notify(AdviceAfterModifyEvent(advice))
+
+        # update item
+        _advice_update_item(item)
 
     # Send mail if relevant
     sendMailIfRelevant(item, 'adviceEdited', 'creators', isSuffix=True)