+++ src/Products/PloneMeeting/utils.py
@@ -2243,40 +2243,6 @@ def _clear_local_roles(obj):
     return old_local_roles
 
 
-def add_wf_history_action(obj, action_name, action_label, user_id=None, insert_index=None):
-    wfTool = api.portal.get_tool('portal_workflow')
-    wfs = wfTool.getWorkflowsFor(obj)
-    if not wfs:
-        return
-    wf = wfs[0]
-    wfName = wf.id
-    # get review_state from last event if any
-    events = list(obj.workflow_history[wfName]) or []
-    if events:
-        previousEvent = events[-1]
-        review_state_id = previousEvent['review_state']
-    else:
-        # use initial_state
-        review_state_id = wf.initial_state
-    # action comments must be translated in the imio.history domain
-    newEvent = {}
-    newEvent['comments'] = action_label or ''
-    # action_name must be translated in the plone domain
-    newEvent['action'] = action_name
-    newEvent['actor'] = user_id or get_current_user_id(obj.REQUEST)
-    # if an insert_index is defined, use same 'time' as previous as
-    # events are sorted on 'time' and just add 1 millisecond
-    if insert_index is not None:
-        newEvent['time'] = events[insert_index]['time'] + 0.000000001
-        newEvent['review_state'] = events[insert_index]['review_state']
-        events.insert(insert_index, newEvent)
-    else:
-        newEvent['time'] = DateTime()
-        newEvent['review_state'] = review_state_id
-        events.insert(len(events), newEvent)
-    obj.workflow_history[wfName] = events
-
-
 def is_editing(cfg):
     """Return True if currently editing something."""
     request = getRequest()